Compile Android tests in build scripts
Android tests aren't compiled in the build script, this causes issues when code changes
are made and the android tests haven't been updated/compiled/run.

While this will not run the android tests on Jenkins it will at least provide
feedback to the author and the reviewers that something is breaking and
probably needs looking into.

Change-Id: I4f2e95eecc306d87080c71364f32987884177367