Dropping minSdkVersion from 19 to 16
To be integrated in some publishers, the SDK should support Android
versions down to 16 during the runtime. However the SDK does not
have to participate to any auction for versions below 19 and may run
in a degraded mode.

This commit does not try to make the actual internal implementation
works for versions below 19, instead it cut the workflow right behind
the public API and always deliver dummy bid responses.
The workflow cut is done because the test coverage quality is very poor
and there is no confidence that actual tests could validate the SDK
for versions below 19.
Moreover, this commit should not change anything in the public API.
Hence, degraded mode and non degraded mode must be transparent.
That's why it is done by abstracting the Criteo class.

This is a merge of fa6cc477289c757613fa91af3d5f43eb39acabeb into
the master branch.

Change-Id: I29aeb48ad6e5dfb78a64f25182811ec629ecb029
JIRA: EE-564