Add unit tests on dummy criteo instance
The dummy instance is used when using the SDK in a degraded manner.
This instance should always answers responses that represent a no bid.

To check the validity, equality should be added on BidToken and
BidResponse. Those are part of the public API, so caution should be
taken when changing them. However, those classes seems to be good
candidates for value objects with equality and hash code overloaded.

Change-Id: Iea18c1059917461bf0c5bb0fce49e0b6b9c0d3e6
JIRA: EE-588