Don't sleep in instrumentation tests
The proposed solution consists in exposing the Executor
implementation used by the AsyncTasks as a dependency
via DependencyProvider. The MockedDependencyRules injects
a CurrentThreadExecutor which forces the execution on the
current thread.

Additionally, because the deviceId is fetched via a callback
mechanism by posting on the main looper, it is necessary to
wait until the MessageQueue associated with the Main Looper
is idle.

JIRA: EE-598
Change-Id: I08bed184ad81386c164cd89c7a85b96a00731088