Test detection of the screen orientation
A dummy activity needs to be created on debug production code side.
Putting it in the test side does not seems to work correctly (see
https://stackoverflow.com/a/51584793 for more information).
This is a safe solution as we should only export release to publishers.
Solving EE-598 will yield a proper solution as the test activities could
then live in a dedicated test module outside of the sdk one.

Change-Id: I4fd9e6df30ad01849bc7efd13cb34435240cbcdf
JIRA: EE-593