E2E for demo to check that display URL is properly in the MoPub webview
Finally, tests on demo creative are abandoned. There is currently an
issue with the demo mode in preprod and PubSDK tests are too weak for
us too afford spending time to fix that yet.
The demo feature is still currently being tested in production by the
PublisherTag, so this is acceptable to take time to mature the PubSDK
project before properly working on validating the demo mode for mobile.
In waiting, this commit still test the demo mode by checking in a E2E
way that the demo display URL is correctly present in the web view of
MoPub and DFP.

Change-Id: I7810a64a526f04b51e706931f5f341d3eba108d4
JIRA: EE-671