Fix DFP interstitial tests in case of no bid from CDB.
Current test synchronization is done thanks to lock based mechanism.
This is well to let dev debug freely tests and be sure that actions
are well finished. However, this may involves some deadlock in case
of expectation that never finished.
Here, tests were waiting for DFP to open an interstitial popup. But,
in case of no bid, no popup appears, meaning a potential deadlock.

Change-Id: Id3b29367a9eae6eb25f1fc1c6ad77835ef9bfe88