Allow to use HTTPS proxy in instrumented test
Below Android N, it's possible to use custom SSL certificates allowing
then to use a HTTPS proxy such as Fiddler or Charles. However, since
Android N, this is not possible unless the application embedded this
network security configuration.

This is setup only for debug artifact, so this should not have any
impact on the release one provided to publishers.

Change-Id: I0ccf93aed61eceace8d3a840cb4760ef9dd5edba