Create only once all provided dependencies
By caching the dependencies, this would let us to factorize internal
dependencies inside the DependencyProvider without any unexpected bugs.
Currently, if A needs B and C needs B, then when writing:

provideA = A(B())
provideC = C(B())

Then we would have different B. As long as they do not have state, this
may be sufficient, but if state is added, then it will be hard to debug.
This is actual the case for DeviceUtil which has the size of the screen
as state and is needed by both Criteo and BidManager. When adding the
BidManager in the DependencyManager, then the DeviceUtil has to be
shared between both.

In order to fix that, the DependencyManager could have another
responsibility: cache the provided dependencies and always return the
first creation. However, this may imply this weird issue:

provideA(x) == provideA(y)

But is this a real issue ? The job of the dependency provider is to give
big abstraction to the SDK. It is not expected that there is multiple
instances of one abstraction.
And this this would be the case, then it may be possible to allow this
by putting the given arguments in the cache keys.

Change-Id: Ib8a95a832080ef69fc59305b92cd16744ba4b45b