Test that getting bid from CDB should timestamp it
Tested scenario:
- Given an ad unit
- Given the clock giving a fixed time
- When getting bid with given ad unit
- And waiting for the bids (idle state)
- Then cache should contain a timestamped bid

This move the moment when the time of download is set for each slots
received from a CDB call. Now, it is set by the bid manager just before
putting them in its bid cache.
This increase the amount of responsibilities of the bid manager, but it
is already handling the TTL, the user-level silent mode. Hence it is
logic that it also handle this. Moreover this reduce the
responsibilities of the Cdb classes which then looks more like a POJO
and represent, then the response from CDB.
Also, this synchronize the moment when the user-level silent mode and
the slot-level silent mode are set.
However, there is a drawback: before the time was set directly after the
network call without change of thread. Now, with the AsyncTasks between,
it is set on the main thread and it may be a delay between the call and
the actual timestamp. But, those AsyncTasks are just a thorn in the
foot. They are just there so the bid manager is always controlled by the
main thread... because it is totally not thread-safe at all.

Change-Id: I62eaa3ebe45b675b66905c2cb27ef19853679cde
JIRA: EE-747