Move validation of displayURL in the slot itself
A lot of callers were checking, double checking, triple checking the
validity of the display URL. This move the validation at a single place:
inside the isValid of slots. Hence every slot return by the bid manager
are totally safe to be used. Also, this simplify a lot of useless
complexity.

Change-Id: Ie055364d6f186c97f35a8381daf86371f7332dad