Test and fix "immediate" bid (CPM > 0, TTL = 0)
Check that getting an immediate bid should populate cache with
15min TTL instead.
Also this move the overriding logic of the TTL into the BidManager, so
the Slot class became only a POJO which responsibility is only to decode
CDB payload.
There is a bug fixed too: TTL is expressed in second, but the constant
used to override the TTL, was set in milliseconds. Hence it was set to
15 000min instead of 15 minutes.

Change-Id: Iadf3e7e93bfd11168d8acc812cee01fba81981b7
JIRA: EE-753
JIRA: EE-765