Test kill switch is persisted when config is updated by remote config
Tested scenario:
- Given remote config call yielding a kill switch status (true / false)
- When SDK is initialized
- And waiting for idle state
- Then new kill switch status is persisted on local storage

This change the actual comportment. Before, the kill switch was persisted
only on new values. This had the side effect that, as the default value
is false (disabled), then on remote call yielding disabled kill switch,
it is not persisted.
This has no downside, except if we want to change the default value in
the future, and that functional description is more complicated and is
not symmetric.

Now the kill switch is always persisted, simplifying the code and the
functional.

Change-Id: I641d91489515c98664c15c7cac7d9e60a98fa263
JIRA: EE-775