Remove the use of lambdas in the SDK
The goal of this change is to make it easier for publishers who are
on minSDK < 26, to integrate the PubSDK.

Below this API level, it is mandatory for the publishers to include in
all their modules build.gradle file:

android {
  compileOptions {
        sourceCompatibility JavaVersion.VERSION_1_8
        targetCompatibility JavaVersion.VERSION_1_8
  }
}

If they don't, desugaring won't work. It is highly probable that most
publishers don't already include these directives.

Since the use of lambdas is facultative, I am erring on the side of
caution by removing lambdas.

Change-Id: Ib4d6479bd2f9450cdb2e10de032cdcf564a9de2d