DeviceInfo return user agent as a future
Getting the user agent is done asynchronously by fetching a webview on
the main thread. In order to be sure that user agent is ready for
everyone using it, it is wrapped into a future. Hence, callers should
wait before using it.
Also, as the future is being completed on main thread, it is not
preferable to fetch it while in the main thread. That's why all calls
have been moved into their respective async tasks.

Change-Id: I1aa0e618d8a97e55c0f51e2adad7edfe34eeee54
JIRA: EE-804