Remove callback on device info init and provide executor to run on UI
Fetching the user agent is synchronized thanks to a future. Hence it's
not necessary to rely on callback to wait for starting the prefetch on
the bid manager.

In order to unit test that, handlers used to run on main thread are
replaced by a provided executor dedicated to allow running on the UI
thread.

Change-Id: I1f8beafa4c9c633c7108b6770d2500ae1959d457
JIRA: EE-804