Update creative served by CDB stub
This has changed because, before, the stub was serving a static JS file
pointing to the previous creative. In order to get creatives resized
accordingly to what is requested, the RTB stub answer now an AJS going
through the CAS stub hosted by CDB. This stub serve then a new creative
that cannot be changed since it's used by PublisherTag tests for its
passback integration.

Change-Id: I38258a8596502dc8f74fa894a2dca622f7be47af