Remove unnecessary destroy method on banner
The destroy method was added for EE-425 to try to fix a memory leak. It
was fixed in the same commit by only using a weak reference of the
banner view instead a keeping a strong one.

Moreover, the destroy method was used to cancel the banner loading task.
But, as detailed in the AsyncTask documentation, the cancel method does
not interrupt any pending task and is only skipping the onPostExecution
method if not already called. In this case, as the async task is only
used to post on the main thread, it's really uncertain that cancelling
the task would have any effect.

Change-Id: Ib9b7fb6f22b833679ae93c3a499d19650e127e64
JIRA: EE-816