Replace AsyncTask for banner into Runnable and executor
The AsyncTask are only used to:
- invoke task on main thread
- post task on a new stack to avoid any locking issues

But the AsyncTask have the downside that it needs to go first on main
thread (for pre execution), then on a background thread (which were
doing nothing in those tasks), then go back on the main thread for post
execution.
This implied technical debt, potential slowness, for just post a task on
main thread for later.

Change-Id: I6b2f8a3992d5950330bf01c767b3a183627069a6
JIRA: EE-607