Encapsulate network URLs in a config class
This decouple the PubSdkApi from the Android context to prepare the
replacement of the configuration by Android string resources by another
solution.

Indeed the Android resources should not be used anymore since it may
not be properly packaged when publishers are wrongly using the AAR as a
JAR. This cause the resources to not be accessible, and then may invoke
some runtime exceptions.

Change-Id: I0dbc247c5b62c3a1473a36376d40c75b0a79f4b5
JIRA: EE-590