Define config URL with gradle properties
The purpose is to empty the android resources files and only rely on the
code or the build. Gradle properties let us handle that at build time.

Indeed the Android resources should not be used anymore since it may
not be properly packaged when publishers are wrongly using the AAR as a
JAR. This cause the resources to not be accessible, and then may invoke
some runtime exceptions.

Change-Id: I56afcfec8c34773171ddd24342fa4d6332025b2a
JIRA: EE-590