Rework bid manager and CDB download task
Split the bid manager responsibility in two components: bid manager and
bid request sender. The new bid manager handle the silence logic and the
bid caching, while the bid request sender handle the dispatching of
requests to the API layer. It keeps references to pending tasks, prevent
fetching twice in parallel the same ad unit, and allow cancellation of
running tasks.
The CDB download task is removed and replaced by future tasks
representing the config call and the cdb call. Both calls are now
completely independent.
This work was made so the bid manager and the bid sender are fully unit
testable to handle corner cases. This also prepare work to handle
collect of CSM.

Change-Id: I68e11374606c75dce7c768f2cc363a14f1e860e9
JIRA: EE-860