Trigger sending of CSM requests
The current strategy of sending metrics is Piggybacking on a request to
CDB because other solutions can bring overcomplexity.

Other solutions that rely on running a periodic background job takes
more time to implement, test and maintain (eg: frequent API changes
between Android versions with different execution guarantees). The
current strategy is to stay away as much as possible from running a
period job.

While we believe that this implementation is acceptable, the risk is
that we receive metrics that are too old (if the user doesn't reopen an
app after a long time).  We should keep an eye on the freshness of the
metrics that will be received on CSM-CDB side.

Change-Id: Ib8be13bf5772d4e54d25a031c31769cc2759caa5
JIRA: EE-984