Implement sending queue factory
This factory use a bounded queue based on the Square's Tape third-party.
It detects if the backing file is stale, and clean it automatically.
In the worst case where it's impossible to have a queue file, it
fallbacks on a in-memory based solution. This very pessimistic
scenario may happens, for instance, if there is no room on device FS.

Change-Id: I3fc533646c804c11643133bbabd23e48e90fb17d
JIRA: EE-985