Persist all remote config response
Previously, only the kill switch was persisted between sessions. This
commit generalize this for the whole response from the remote config
TLA. This is needed because there is no synchronization between the CDB
call and the remote config call, and they are done in parallel. Without
persistence, a bid may be used before the remote config end, and so
default value would be used instead of remote config. This is acceptable
for a first call, but we should at least try to be better on the next
sessions by persisting the config.

Change-Id: I4ccc71935484242423c0b550ea2ef65a7377e492
JIRA: EE-1009