Call onAdClosed when user come back from an ad
On banner ad, when user click on the banner, he gets redirected to the
advertiser website. When he goes back to the application (by pressing
the go back button or by an other way), the onAdClosed callback is
triggered.
On interstitial ad, when user click on the close button, the onAdClosed
callback is triggered (this was already handled). And the same logic
than banner should happen there.

Test on interstitial will be done, but unfortunately, this is currently
not possible on banner, since it is not currently possible to mock a CAS
server (adding a dependency such as mock-server make the Android test
APK explodes the 64K limit, and EE-589 should be tackled before).

Change-Id: I777513fd1a664ca210e80f29856e61cb47c4aa78
JIRA: EE-879