Validate onAdClose callback on interstitial
This validates that the callback is invoked when the user click on the
ad and go back. This is tested via deeplink into an internal activity.
Unfortunately, it is not possible to test this when having a deeplink to
an external application.
But this case was tested manually and seems to react like when using an
internal activity.

Change-Id: I17d7390b6642bfe218c1c9e0ce9f31bd8b1353c6
JIRA: EE-879