Provide small overlay above the Android log API
The SDK does not require a lot of features on logging: just a global
switch to deactivate debug logging. Although, we should not deactivate
any logging on publisher's applications.
The Android logging is too simple and do not allow such filtering. The
J.U.L (java.util.logging) provide this kind of filtering and is
already provided by the JDK. But it is quite big and we may not need
such powerful features. Others are external dependencies or/and may be
too big (SLF4J, log4j, ...).

Change-Id: Ia6875819e92226ae31d8fc9f7c50de89a05ae810
JIRA: EE-688