Fix notification of listener for Android version < 19
The SDK run in a degraded mode before version 19. It should not crash,
is not obliged to bid, but should at least work as a no bid provider. So
when fetching a bid, listeners should be notified for a no fill.
The notification of banner listener was broken in standalone. The
incident may be minimal for standalone, but this also implies that this
bug also appears on mediation adapters. They implicitly accept any
android version since the public documentation shows that setup could be
done by copy/pasting the mediation code.

Change-Id: Ie3c339fef69fb2abeebaf2d68ef6c5994102aa86
JIRA: EE-1020