Remove dependency to AtomicFile from Android support
AtomicFile are used by the CSM feature in the metric folder. AtomicFile
from Android support allow us to use this for any android versions. But
this brings the android support lib in the project. This is a lib that
should match the version of the build tools used to build the project.
If a publisher import the SDK and is using a different version, there
may be some collisions.
To reduce this risk, we can remove this part. This have the downside to
remove the CSM feature for API version below 17. But, SDK is already
degraded below version 19.

Change-Id: Id667f1319df63eeedc8bd78c186196bf88c6d748
JIRA: EE-1024