Remove dependency to Android support
Android support is only use for JSR-305 annotations. So we can just
keep the dependency at compile time. If kept at runtime, this implies
issues with publisher integrations. This is a lib that should match the
version of the build tools used to build the project. If a publisher
import the SDK and is using a different version, there
may be some collisions.

Change-Id: I1790c76d929f49a5d9b076e91fe4cd36be60927e
JIRA: EE-1024