Add adapters as dependency of the test app
Since the test app is still inside the SDK project, this looks like a
loop. But we only have yet the test app acting as a gradle integration
test. So this is better than nothing to detect HUGE bug in artifacts
(such as zipped empty aar file).
Moreover, this allows to use the test activities related to mediation
for manual testing.

Change-Id: Ib01f98cf702527d5131d024205693227a7622436