Rework the serialization and data model of native assets
The previous classes do not identify at all the nullability or the
emptiness of data in native assets. Moreover some checks are useless
regarding the CDB data layer.
The new one considers that all elements are not nullable (in fact they
are empty if no value is given by RTB). At least one product and one
impression pixel should be present in the native payload. URLs are
treated directly as URL. This implies that, once constructed, the object
is already considered as valid and no further validation step is
required.

Change-Id: I8bead6c79d72e08acf9944d8ec3af492b69a0707
JIRA: EE-1039