Fix expectations about running in an executor
The expectations done with DirectMockRunOnUiThreadExecutor were broken
because some production code use SafeRunnable that catch exceptions.
So some tests were not really checking anything. To let production code
catch whatever it wants, and to avoid tests having to know this, it is
fixed by decoupling the expectation step from the verification step.
Expectation registers potential failures, and verification throws them.
As verification is done in test code, outside the SafeRunnable, failures
will be reported as expected.

Change-Id: I287084cc1347895b9ca601329287b392a60f8867