Add API and test cases for recycler views
This add 2 new public methods in the API:
- CriteoNativeAd#renderNativeView
- CriteoNativeLoader#createEmptyNativeView

The first one allow publishers to render many times their native ad on
different views. This is mandatory to handle recycler views.

The second one is mandatory for RecyclerView, so publisher can create
their empty native view in the onCreateViewHolder method without having
any bid. They should call this method rather than the create method in
the renderer, so AdChoice icon is injected internally.

Functional tests are extended to test the new API in the same manner
than for ad layout, and for both Standalone and In-House.

Change-Id: Id2b383fe02ec8e34402f55a11d9f5f465b76b9d3
JIRA: EE-1068