Create a new module dedicated to Android tests
We reach the limit of the method count allowed in a single DEX file (64K
methods). On instrumented tests, proguard or R8 is not run correctly, so
dependencies take a lot of methods.
MultiDex is a solution offered by Android SDK. But it cannot be applied
only on test, but should be applied on a whole module.
That's why instrumented tests will be moved in a dedicated module out
the SDK one.

Change-Id: I9e80a085df76fd57263a09b6d179b4b17fca3e4d
JIRA: EE-589