Move activity accepting criteo-test:// deeplink
Mediation adapter rely on test-utils and when running their test, a test
APK with this activity is installed on the device.
So 3 applications (SDK and mediation tests) are accepting this deeplink.
This make test using it hang because device need an input to select an
application.
If mediation adapters need those tests one day, they would need their
own deeplink to avoid such pain.

Change-Id: Ic3afb5af6938ccbb0aa20d09d68f4860fef1a353
JIRA: EE-589