Migrate to AndroidX
Android support is deprecated. While there is no urgent
need to upgrade to AndroidX, it is still recommended to do so.

This will allow us to upgrade MoPub's dependencies in the
test application (and potentially in the adapters, once they will be
migrated as well).

With this change:
- publishers that already use AndroidX won't have to add:
android.useAndroidX = true and android.enabledJetifier = true because of
us.

- publishers that don't use AndroiX might have issues when building
their project, but there are workarounds: they can
either migrate to AndroidX or exclude the AndroidX artifacts.

This change also enables multidex support on the test app, as the 64K
limit was reached.

Change-Id: I162cf1d6025af93ba7a5c6438c9036c301625197
JIRA: EE-830