Update Slack plugin version to 0.2.0
This version change the default Git block format by providing a "git
describe" of the HEAD instead of the current branch name. With the
current CI workflow, HEAD is detached from the master branch (or any
other branch) so the branch name is the commit SHA-1, which is
misleading when we're cherry-picking commits across branches. With git
describe, the nearest refs (branch or tag) of HEAD is retrieved and
displayed.

Change-Id: I281ce4fe8638a296bbf49a741ab38717320632fe