Speed up bearcat functional tests
The BearcatPrivacyFunctionalTest is a parametrized tests with a lot of
parametrization. Moreover, for each test cases, a new activity is opened
then closed. The test spend almost all its time in this.
This manually trigger the AppEvents without starting an activity. This
make the tests run in 10s (more than one minute before).
Other tests are already tested the trigger of the AppEvents such as
CriteoFunctionalTest or BearcatFunctionalTest.

Change-Id: Ifcaf63d5e3e0249059a55ab5ec8b718e527e45cf