Add a constructor of AdChoiceOverlayNativeRenderer for MoPub adapter
The MoPub adapter cannot access the dependency provider, but for native,
it needs the access to the ad choice renderer. So SDK offers an easy to
use constructor.

This constructor eagerly access internal dependencies and make the
integration crash if the SDK is not initialized and force the publisher
to implement this step. By any chance, this may improve the performance
of the mediation as publishers would now have to prefetch the bids.

Change-Id: I2389ed9f27821971eeb616f80086f3806ac41ad5
JIRA: EE-1107