Provide setRenderer method for MoPub mediation adapter
In MoPub mediation, publisher cannot access directly the NativeLoader.
Instead, they need to register a renderer outside the loader flow.
To accommodate the adapter, the CriteoNativeAd provide a setRenderer, so
that the adapter can swap it with the registered one just before the
actual rendering step.

Note that it is not possible to @RestrictTo the method, as the group ID
of the SDK is not the same than the group ID of the adapter. But the
method is package-private, and is not sensitive.

Change-Id: Ie688587f263554b5cf59a5c60d9edd7723c13c60
JIRA: EE-1107