Add profileId in remote config request
This field is not used by the old remote config TLA. But the new one
would need it. This is done to handle the work that will be done in Q3:
having dedicated profile ID per integration. This would allow us to have
a finer config by integration (for instance, if a major bug is found on
a integration, the kill switch can be activated).

Change-Id: If0aba0da3c98b469dd13c4965be3764fc39f5c15
JIRA: EE-1048