Implement a CDB mock handling CSM request
The mock is not applied for now, but an integration tests reconciles the
outcome of this mock compared to the true CDB preprod.

To ease the development, the integration tests are put in this git
project. But when the project will be open source and tests run on a
public CI, those tests should be extracted in order to keep the access
to the preprod network.

Change-Id: I468c71eef54ac711036f919079db188e9df969fb
JIRA: EE-1137