Automatically recreate CSM queue file if it gets corrupted
This just happens to me randomly: an empty byte array was stored inside
the queue file. Those issues didn't seem to be fully recoverable:
- the current converter does not allow empty byte array
- after removing this empty byte array, things may get worse and the
queue file could really be broken (allocating byte array with negative
length)

This destroys and recreates the queue file if it detects that it is
corrupted.

Change-Id: I861a57cc206948dcb83eca23800be83259967b90