Support few TCF consent strings in CDB mock for E2E tests
This only denies few consent strings of TCF1/TCF2 in the CDB mock. A
proper implementation should decode the string, check the consent,
handle TCF1 and TCF2. This is FAR, and we can imagine that TCF constant
will not change and that this is sufficient for a mock server.

Change-Id: Ie4fd446e480da5b7c6de9ed8e4a4d34ba4fc6dcf
JIRA: EE-1137