Deploy the test app in a dedicated package on Bintray
JCenter only accept packages representing library with .jar or .aar
The test app is not in this case, and there is no meaning to sync it on
JCenter anyway. Hence, we push it on Bintray but in a dedicated package.
So the SDK and the test utils can be sync with JCenter.

JIRA: EE-1167