Synchronize image loading with waitForIdleState
Picasso was not synchronized through the waitForIdleState. Hence it was
not possible to reliably wait for downloaded images. Sleeps were added,
but it doesn't works well.
This makes appear a notion of `AsyncResources` which can help to declare
pending async tasks based on listener callbacks, such as Picasso.
This is a no-op on production.