Fetch device info asynchronously on the UI thread
Some publishers detect that the SDK is taking a long time during init.
This is mainly due to the user-agent retrieval that needs access to the
UI thread to open a webview and get the UA in there.
To make the publisher's UI smoother, the UA job is still on UI thread
but is posted asynchronously. It is expected that SDK init is faster
from an UX point of view making their applications start quicker.

On the other hand, our bidding mechanism needs the UA and will be
delayed via concurrent sync. It is expected to see a increased delay of
50ms on recent devices, and up to 500ms on old devices.

JIRA: GSTM-13350