Directly call AdvertisingIdClient methods
Those methods are in the compile classpath and are expected to be
provided in the runtime classpath by publishers as explained in the
public documentation. Hence reflection can be removed.

Logging in case of error is improved:
- if publishers didn't brings the good dependency, they are notified to
- sometimes, in tests, it is impossible to read the advertising ID and
there is no clue of why. The stacktrace will now be printed in logcat.