Stabilize CSM test about timeout network
The logcat shows that, even by setting a really small timeout threshold,
timeout are not always throws by the network layer. This may be due
because the CDB mock server is embedded and fast.

In order to stabilize, the test, in addition of the small timeout
threshold, the CBB mock server is also throttled.