Use fixed version for the upload-artifact GA
This GA is now using "main" as default branch. Using the master one
yield this warning:

The "master" branch is no longer the default branch name for
actions/upload-artifact
Please pin to a specific version or use "main".

To stay out of any change a fixed version is used instead.