Remove sleep in test to reduce flakyness
This test had failed because there was two metrics read. This means that
the read task occurred after the second update task. This was guarded
via a sleep that is not deterministic, hence the flakyness.

This fix this flakyness by forcing the execution: second update is
blocked until read is finished.