Use current size of screen for interstitial
Before, we had the concept of portrait and landscape sizes and we were
detecting the orientation to know which one to use.
Moreover, the portrait size was defined as the application size during
Criteo init, and the landscape size was defined as the transposed of the
portrait.

There was issue with this:
- EE-852: when starting an application in landscape, the previous
implementation of the orientation detection returned the wrong
orientation. This yields interstitial that were always in the wrong
direction.

The new strategy does not bother using concept such as orientation,
portrait or landscape. Instead, the current size of the screen is just
used as the size wanted for interstitial. So there is no more
assumptions on the relation between landscape and portrait size, and no
more data retention making use of wrong size.

The orientation is still used at one place: for DFP header bidding when
special interstitial sizes should be injected depending of the current
screen orientation. However, the orientation implementation is fixed.

JIRA: EE-852