Move back the android tests in the publisher-sdk module
Initially, android tests were moved in a dedicated modules because we
reached the limit of 64K methods in a single DEX file. It is possible to
enable the multidex, but it should be applied on the whole module,
production code included. So tests were extracted to not have any impact
on publishers.

However, mediation adapters show that it is possible to just remove the
multidex from the exported dependencies, solving the original issue.

Moreover having the android tests in a dedicated modules make the
project harder to maintain:
- IDE are not fully integrated with such structure
- overall project is more complex
- code coverage does not at all

On the last point, some experiences were made to get code coverage
working across module. Unfortunately, coverage offered by AGP is not
configurable at all and only more legacy could maybe have chance to make
this work during a few times only.

Note that the publisher-sdk-tests module may still be useful. Dummies
activities used for tests can lives there without impacting production
code.