Run all tests on merge, but only flaky ones are retried
Only tests declared as flaky are run with a retry policy. Other tests
are always run normally.
This prevents any true negative to pass the tests.
Also, the Gordon runner does not support Jacoco coverage. So only flaky
tests are not counted in the coverage report. Which is a good thing,
because we cannot affirm that something is really covered if the tests
are flaky.
Moreover, runners with retry policies, such as Gordon, are slow compared
to the default test runner. This would speed up the CI on merge.