Add option to mock logger via the MockedDependenciesRule
When using injection annotations, such as @Inject, it is not possible to
interact with test before creation and injection of components. In the
case of the logger, it is generally instantiated at component creation,
hence it is not mockable with injection annotations.
This option makes the logger and logger factory managed by the mocked
dependency provider and solves the above issue.