Wait uninterruptibly when running task on UI thread from tests
The methods runOnMainThreadAndWait and callOnMainThreadAndWait are
written in a way, so that the callers can write code that is
sequentially executed from top to bottom without any concurrency even if
the code in those methods is running on another thread.
Hence, in this case it is not correct to be interrupted by someone as it
would remind about a concurrency dimension that should be forgot here.