Add synchronization on flaky test
The "whenLoadingAnInterstitial GivenBidAvailableTwice
DisplayUrlIsProperlyLoadedInInterstitialActivityTwice" in Standalone
tests is very very flaky and makes the build fail even after 5 retries.

The assertion shows that the webview does not seems to contains any HTML
yet. Although, the listeners are correctly notified and the interstitial
is marked as being loaded. This means that the creative is downloaded,
but not yet loaded in the webview.

There were already few cases before where a webview takes time to be
fully loaded. So we can hope that waiting for the webview content will
help to make this test pass.