Use DummyCriteo if not initialized properly (#89)
* Use DummyCriteo if not initialized properly

Problem:
When an error happens during the initialization path, Criteo instance
will be `null` and any future call to Criteo.getInstance() will throw a
RuntimeException.

Solution:
DummyCriteo is assigned to Criteo singleton object, when any runtime
exception is thrown during initialization. While the integration will
not work, it will prevent the host application from crashing. Logging
will allow us to investigate and correct the root cause with the
publisher.

* Guard against null application in CriteoInterstitial

Problem:
An `application` object is required for the SDK to operate properly.
However, nothing prevents CriteoInterstitialActivity from being open even if
`application` is null.

Solution:
CriteoInterstitial APIs must check wehther `application` is null or
return immediately. This is a workaround to prevent apps from crashing
in production but actions need to be taken in the future to:
1. Understand how `application` can be null for publishers.
2. Check whether we can reduce the scope of `application` within the SDk
and use an Activity context when appropriate (in this case, the Activity
context that is passed to CriteoInterstitial could be used).
3. Understand the implications of making the application object
nullable.